Fixed class inconsistencies for multilure
