+++ Localization/en-US.hjson
@@ -2,82 +2,21 @@ Mods: {
 	BetterFishing: {
 		Configs: {
 			MultilureMode: {
-				Tooltip: Changes how the multilure complexity and conditions works
-				DISABLED.Label: DISABLED
+				Tooltip:
+					'''
+					Changes how the multilure complexity and conditions works
+					[c/42b9f5:Disabled]: Fully disables the multilure feature
+					[c/42b9f5:Simple]: There is a fixed amount of lines per rod
+					[c/42b9f5:Complex]: The amount of lines depends on rod conditions such as biome or time of day
+					'''
+				DISABLED.Label: "[c/cc5343:DISABLED]"
 				SIMPLE.Label: SIMPLE
-				NORMAL.Label: NORMAL
+				NORMAL.Label: "[c/80e0b0:COMPLEX]"
 				Label: Multilure Mode
 			}
 
 			Header.Multilure: "[i:2373] Multilure"
 			ConfigName: Global configuration
-		}
-
-		Multilure: {
-			Vanilla: {
-				SIMPLE: {
-					WoodFishingPole: This rod fires {0} {^0:line;lines at the same time}
-					ReinforcedFishingPole: This rod fires {0} {^0:line;lines at the same time}
-					FisherofSouls: This rod fires {0} {^0:line;lines at the same time}
-					Fleshcatcher: This rod fires {0} {^0:line;lines at the same time}
-					ScarabFishingRod: This rod fires {0} {^0:line;lines at the same time}
-					BloodFishingRod: This rod fires {0} {^0:line;lines at the same time}
-					GoldenFishingRod: This rod fires {0} {^0:line;lines at the same time}
-					HotlineFishingHook: This rod fires {0} {^0:line;lines at the same time}
-					SittingDucksFishingRod: This rod fires between {0} and {1} {^1:line;lines}
-					MechanicsRod: This rod fires between {0} and {1} {^1:line;lines}
-					FiberglassFishingPole: This rod fires {0} {^0:line;lines at the same time}
-				}
-
-				NORMAL: {
-					WoodFishingPole: This rod fires {0} {^0:line;lines at the same time}
-					ReinforcedFishingPole: This rod fires {0} {^0:line;lines at the same time}
-					FisherofSouls: This rod fires {0} {^0:line;lines} plus {1} extra at Corruption
-					Fleshcatcher: This rod fires {0} {^0:line;lines} plus {1} extra at Crimson
-					ScarabFishingRod: This rod fires {0} {^0:line;lines} and consumes {2} mana per extra line (up to {3} lines)
-					BloodFishingRod: This rod fires {0} {^0:line;lines} and consumes {2} health per extra line (up to {3} lines)
-					GoldenFishingRod: This rod fires {0} {^0:line;lines} and has a {2}% chance to shoot {1} extra {^1:line;lines} and a {4}% chance to shoot {3} extra {^3:line;lines}
-					HotlineFishingHook: This rod fires between {0} and {1} {^1:line;lines} depending on the heat of the biome
-					SittingDucksFishingRod: This rod fires between {0} and {1} {^1:line;lines}
-					MechanicsRod: This rod fires between {0} and {1} {^1:line;lines} plus {2} extra {^2:line;lines} at night
-					FiberglassFishingPole: This rod fires between {0} and {1} {^1:line;lines} depending on the humidity of the biome
-				}
-			}
-
-			CalamityMod: {
-				NORMAL: {
-					WulfrumRod: This rod fires {0} {^0:line;lines at the same time}
-					SlurperPole: Throws between {0} and {1} {^1:line;lines} depending on the heat of the biome. Doesn't work on cold biomes
-					HeronRod: This rod fires {0} {^0:line;lines} and adds up to {1} depending speed of the wind
-					VerstaltiteFishingRod: Throws between {0} and {1} {^1:line;lines} depending on the heat of the biome. Doesn't work on hot biomes
-					RiftReeler: This rod fires between {0} and {1} {^1:line;lines} and can add up to {2} lines depending of the heat of the biome
-					FeralDoubleRod: This rod fires {0} {^0:line;lines at the same time}
-					EarlyBloomRod: This rod fires {0} {^0:line;lines at the same time} and adds {1} {^1:line;lines} at Jungle
-					NavyFishingRod: This rod fires between {0} and {1} {^1:line;lines}
-					TheDevourerofCods: This rod fires {0} {^0:line;lines at the same time}
-				}
-
-				SIMPLE: {
-					WulfrumRod: This rod fires {0} {^0:line;lines at the same time}
-					SlurperPole: This rod fires {0} {^0:line;lines at the same time}
-					HeronRod: This rod fires {0} {^0:line;lines at the same time}
-					VerstaltiteFishingRod: This rod fires {0} {^0:line;lines at the same time}
-					RiftReeler: This rod fires between {0} and {1} {^1:line;lines}
-					FeralDoubleRod: This rod fires {0} {^0:line;lines at the same time}
-					EarlyBloomRod: This rod fires {0} {^0:line;lines at the same time}
-					NavyFishingRod: This rod fires between {0} and {1} {^1:line;lines}
-					TheDevourerofCods: This rod fires {0} {^0:line;lines at the same time}
-				}
-
-				REPLACE.TheDevourerofCods: Line never snaps and can fish in lava
-			}
-		}
-
-		Buffs: {
-			AumentedFishingPower: {
-				DisplayName: Fishing inspiration
-				Description: Increments your fishing power
-			}
 		}
 	}
 }
