+++ Multilure/Multilure.cs
@@ -1,7 +1,7 @@
 ﻿using BetterFishing.Config.Model;
-using BetterFishing.Multilure.PostThrow;
 using System;
 using System.Collections.Generic;
+using Terraria.ModLoader;
 
 namespace BetterFishing.Multilure
 {
@@ -14,7 +14,6 @@ namespace BetterFishing.Multilure
         public const string DESCRIPTION_PREFIX = "[i:2373] ";
         public const string TOOLTIP_DEFINITION = "BetterFishing";
 
-        private static Dictionary<MultilureMode, Dictionary<int, MultilurePostThrow>> _postThrows;
         private static Dictionary<MultilureMode, Dictionary<int, MultilureLine[]>> _lines;
         private static Dictionary<MultilureMode, Dictionary<int, MultilureDescription[]>> _descriptions;
 
@@ -25,24 +24,12 @@ namespace BetterFishing.Multilure
 
             _lines = new(size);
             _descriptions = new(size);
-            _postThrows = new(size);
 
             foreach (MultilureMode mode in modes)
             {
                 _lines.Add(mode, new Dictionary<int, MultilureLine[]>());
                 _descriptions.Add(mode, new Dictionary<int, MultilureDescription[]>());
-                _postThrows.Add(mode, new Dictionary<int, MultilurePostThrow>());
             }
-        }
-
-        internal static void SetPostThrow(MultilureMode mode, int itemId, MultilurePostThrow postThrow)
-        {
-            _postThrows[mode].Add(itemId, postThrow);
-        }
-
-        internal static MultilurePostThrow GetPostThrow(int itemId)
-        {
-            return _postThrows[BetterFishing.Configuration.MultilureMode].GetValueOrDefault(itemId, null);
         }
 
         internal static void SetLines(MultilureMode mode, int itemId, List<MultilureLine> line)
