+++ changelog.txt
@@ -1,1 +1,4 @@
 ﻿Version 2.2.0
+
+[h2]Fixed[/h2]
+-Minor code inconsistecies for multilure module
